removing get_streams from instrument and json writer
